+++ sparse_dot_mkl/sparse_dot.py
@@ -4,9 +4,10 @@ from sparse_dot_mkl._dense_dense import _dense_dot_dense as _ddd
 from sparse_dot_mkl._sparse_vector import _sparse_dot_vector as _sdv
 from sparse_dot_mkl._gram_matrix import _gram_matrix as _gm
 from sparse_dot_mkl._sparse_qr_solver import sparse_qr_solver as _qrs
-from sparse_dot_mkl._mkl_interface import get_version_string, _is_dense_vector
+from sparse_dot_mkl._mkl_interface import print_mkl_debug, _is_dense_vector, set_debug_mode, get_version_string
 import scipy.sparse as _spsparse
 import numpy as _np
+import warnings
 
 
 def dot_product_mkl(matrix_a, matrix_b, cast=False, copy=True, reorder_output=False, dense=False, debug=False,
@@ -34,7 +35,7 @@ def dot_product_mkl(matrix_a, matrix_b, cast=False, copy=True, reorder_output=Fa
     This does not require any copy and is memory efficient if the output array density is > 50%
     Note that this flag has no effect if one input array is dense; then the output will always be dense
     :type dense: bool
-    :param debug: Should debug and timing messages be printed. Defaults to false.
+    :param debug: Deprecated debug flag. Use `sparse_dot_mkl.set_debug_mode(True)`
     :type debug: bool
     :param out: Add the dot product to this array if provided.
     :type out: np.ndarray, None
@@ -44,12 +45,8 @@ def dot_product_mkl(matrix_a, matrix_b, cast=False, copy=True, reorder_output=Fa
     :rtype: scipy.sparse.csr_matrix, scipy.sparse.csc_matrix, np.ndarray
     """
 
-    dprint = print if debug else lambda *x: x
-
-    if get_version_string() is None and debug:
-        dprint("mkl-service must be installed to get full debug messaging")
-    elif debug:
-        dprint(get_version_string())
+    warnings.warn("Set debug mode with sparse_dot_mkl.set_debug_mode(True)", DeprecationWarning) if debug else None
+    print_mkl_debug()
 
     num_sparse = sum((_spsparse.issparse(matrix_a), _spsparse.issparse(matrix_b)))
 
@@ -58,19 +55,19 @@ def dot_product_mkl(matrix_a, matrix_b, cast=False, copy=True, reorder_output=Fa
         raise ValueError("out argument cannot be used with sparse (dot) sparse matrix multiplication")
 
     elif num_sparse == 2:
-        return _sds(matrix_a, matrix_b, cast=cast, reorder_output=reorder_output, dense=dense, dprint=dprint)
+        return _sds(matrix_a, matrix_b, cast=cast, reorder_output=reorder_output, dense=dense)
 
     # SPARSE (DOT) VECTOR #
     elif num_sparse == 1 and _is_dense_vector(matrix_a) and (matrix_a.ndim == 1 or matrix_a.shape[0] == 1):
-        return _sdv(matrix_a, matrix_b, cast=cast, dprint=dprint, out=out, out_scalar=out_scalar)
+        return _sdv(matrix_a, matrix_b, cast=cast, out=out, out_scalar=out_scalar)
 
     # SPARSE (DOT) VECTOR #
     elif num_sparse == 1 and _is_dense_vector(matrix_b) and (matrix_b.ndim == 1 or matrix_b.shape[1] == 1):
-        return _sdv(matrix_a, matrix_b, cast=cast, dprint=dprint, out=out, out_scalar=out_scalar)
+        return _sdv(matrix_a, matrix_b, cast=cast, out=out, out_scalar=out_scalar)
 
     # SPARSE (DOT) DENSE & DENSE (DOT) SPARSE #
     elif num_sparse == 1:
-        return _sdd(matrix_a, matrix_b, cast=cast, dprint=dprint, out=out, out_scalar=out_scalar)
+        return _sdd(matrix_a, matrix_b, cast=cast, out=out, out_scalar=out_scalar)
 
     # SPECIAL CASE OF VECTOR (DOT) VECTOR #
     # THIS IS JUST EASIER THAN GETTING THIS EDGE CONDITION RIGHT IN MKL #
@@ -81,7 +78,7 @@ def dot_product_mkl(matrix_a, matrix_b, cast=False, copy=True, reorder_output=Fa
 
     # DENSE (DOT) DENSE
     else:
-        return _ddd(matrix_a, matrix_b, cast=cast, dprint=dprint, out=out, out_scalar=out_scalar)
+        return _ddd(matrix_a, matrix_b, cast=cast, out=out, out_scalar=out_scalar)
 
 
 def gram_matrix_mkl(matrix, transpose=False, cast=False, dense=False, debug=False, reorder_output=False,
@@ -100,7 +97,7 @@ def gram_matrix_mkl(matrix, transpose=False, cast=False, dense=False, debug=Fals
     :type cast: bool
     :param dense: Produce a dense matrix output instead of a sparse matrix
     :type dense: bool
-    :param debug: Should debug and timing messages be printed. Defaults to false.
+    :param debug: Deprecated debug flag. Use `sparse_dot_mkl.set_debug_mode(True)`
     :type debug: bool
     :param reorder_output: Should the array indices be reordered using MKL
     The scipy sparse dot product does not yield ordered column indices so this defaults to False
@@ -111,13 +108,9 @@ def gram_matrix_mkl(matrix, transpose=False, cast=False, dense=False, debug=Fals
     :type out_scalar: float, None
     :return: Gram matrix
     :rtype: scipy.sparse.csr_matrix, np.ndarray"""
-    
-    dprint = print if debug else lambda *x: x
 
-    if get_version_string() is None and debug:
-        dprint("mkl-service must be installed to get full debug messaging")
-    elif debug:
-        dprint(get_version_string())
+    warnings.warn("Set debug mode with sparse_dot_mkl.set_debug_mode(True)", DeprecationWarning) if debug else None
+    print_mkl_debug()
 
     return _gm(matrix, transpose=transpose, cast=cast, dense=dense, reorder_output=reorder_output,
                out=out, out_scalar=out_scalar)
@@ -135,20 +128,16 @@ def sparse_qr_solve_mkl(matrix_a, matrix_b, cast=False, debug=False):
     and should a CSR matrix be cast to a CSC matrix.
     Defaults to False
     :type cast: bool
-    :param debug: Should debug messages be printed. Defaults to false.
+    :param debug: Deprecated debug flag. Use `sparse_dot_mkl.set_debug_mode(True)`
     :type debug: bool
     :return: Dense array X
     :rtype: np.ndarray
     """
-    
-    dprint = print if debug else lambda *x: x
-    
-    if get_version_string() is None and debug:
-        dprint("mkl-service must be installed to get full debug messaging")
-    elif debug:
-        dprint(get_version_string())
-        
-    return _qrs(matrix_a, matrix_b, cast=cast, dprint=dprint)
+
+    warnings.warn("Set debug mode with sparse_dot_mkl.set_debug_mode(True)", DeprecationWarning) if debug else None
+    print_mkl_debug()
+
+    return _qrs(matrix_a, matrix_b, cast=cast)
 
   
 # Alias for backwards compatibility